Check Your Study
This project is designed for students preparing for university entrance exams. It calculates the daily number of questions that need to be solved based on their net scores in practice tests. Additionally, it provides recommendations for supplementary resources and suggests potential universities. The aim is to assist students who have limited access to guidance counselors.